Added Shooting Enemies and Player

+++ Assets/App/Scripts/Game/Entity/Health/IDamageable.cs
@@ -2,7 +2,7 @@
 {
     public interface IDamageable
     {
-        public void TakeDamage(int damage);
+        public void TakeDamage(float damage);
 
         public void Kill();
     }
